refactor: Answer 도메인 test package나누기
NullCheck인터페이스 생성
