+++ src/main/java/qna/action/NullCheckAction.java
@@ -1,0 +1,13 @@
+package qna.action;
+
+import qna.NotFoundException;
+
+import java.util.Objects;
+
+public interface NullCheckAction {
+    default void throwExceptionIsNullObject(Object object) {
+        if (Objects.isNull(object)) {
+            throw new NotFoundException();
+        }
+    }
+}
