+++ src/test/java/qna/answer/AnswerRepositoryTest.java
@@ -1,12 +1,10 @@
-package qna.domain;
+package qna.answer;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import qna.answer.Answer;
-import qna.answer.AnswerRepository;
 import qna.question.Question;
 import qna.question.QuestionRepository;
 import qna.user.User;
